Add a thresholded regex and adjacency table predictor
The regex & adjacency table predictors tend to get "diluted" by having many
different headers expressed by the same regex & set of adjacency tables.

Resist this effect by having a similarity threshold that applies to the
total number of characters in the candidate header as a percentage of the
total number of explicitly matched characters in the regex.