Increase move confidence
If there's low confidence the message should remain in the inbox.

If there's 2 highly likely candidate destinations for the message outside
the inbox, the algorithm is hesitant to move messages even if the inbox has
0 confidence that the message should reside there.

In this case, the best option is likely still to choose the box with the
highest confidence.