Sync & Store Messages
There's no very reliable way to ID messages when they move between
folders. Some servers support globally unique IDs. For now, compromising
with a hash of the sender header(s) and the sent & received dates.

Maybe just support servers with GUIDs?

Remember the latest message we've synced for each mailbox.

Create the synchronizer the same way in the test as in the app.